MS-59 fix(bugs): fix bugs related to login and logout
