+++ src/api/api.ts
@@ -1,7 +1,8 @@
-import changeAppAfterLogout from "src/pages/loginPage/logoutHandler";
+import { Pages } from "src/types/types";
+import { getItemFromLocalStorage } from "src/utils/utils";
+import { menuItemLogIn, menuItemLogOut, menuItemSingUp, userMenu, addUserGreetingToHeader } from "src/pages/basePage/basePage";
 import { createApiBuilderFromCtpClient, MyCustomerDraft } from "@commercetools/platform-sdk";
 import { state } from "src/store/state";
-import { addUserGreetingToHeader } from "src/pages/basePage/basePage";
 import generateAnonymousSessionFlow from "./anonymousClientBuilder";
 import generateRefreshTokenFlow from "./refreshTokenClientBuilder";
 
@@ -10,7 +11,11 @@ if (!state.refreshToken) {
   ctpClient = generateAnonymousSessionFlow();
 }
 if (state.refreshToken) {
-  changeAppAfterLogout();
+  window.location.hash = Pages.MAIN;
+  menuItemLogIn.href = Pages.MAIN;
+  menuItemSingUp.href = Pages.MAIN;
+  state.name = getItemFromLocalStorage<string>("user");
+  userMenu.append(menuItemLogOut);
   addUserGreetingToHeader();
   ctpClient = generateRefreshTokenFlow();
   const apiRoot = createApiBuilderFromCtpClient(ctpClient).withProjectKey({ projectKey: "steps-moon-store" });
