+++ src/store/state.ts
@@ -2,8 +2,8 @@ import { getItemFromLocalStorage } from "src/utils/utils";
 import { State } from "../types/types";
 
 export const state: State = {
-  name: getItemFromLocalStorage("user"),
-  refreshToken: getItemFromLocalStorage("refreshToken"),
+  name: getItemFromLocalStorage<string>("user"),
+  refreshToken: getItemFromLocalStorage<string>("refreshToken"),
 };
 
 export default state;
