+++ src/utils/utils.ts
@@ -1,15 +1,33 @@
 export const priceFormatter = new Intl.NumberFormat("ru-RU", { style: "currency", currency: "RUB", minimumFractionDigits: 0 });
 
-export function setItemToLocalStorage(key: string, value: string | undefined) {
-  localStorage.setItem(key, JSON.stringify(value));
+// export function setItemToLocalStorage(key: string, value: string) {
+//   localStorage.setItem(key, value);
+// }
+
+// export function getItemFromLocalStorage(key: string) {
+//   if (localStorage.getItem(key)) {
+//     const item = localStorage.getItem(key);
+//     return item;
+//   }
+//   return null;
+// }
+
+export function setItemToLocalStorage<T>(key: string, value: T) {
+  localStorage.setItem(key, JSON.stringify({ value }));
 }
 
-export function getItemFromLocalStorage(key: string): string | null {
+export function getItemFromLocalStorage<T>(key: string): T | string | null {
   const item = localStorage.getItem(key);
-  if (item) {
-    return JSON.parse(item);
+
+  if (!item) {
+    return null;
   }
-  return null;
+
+  try {
+    return JSON.parse(item).value;
+  } catch {
+    return item;
+  }
 }
 
 export default priceFormatter;
